+++ src/Core/Translate.hpp
@@ -10,7 +10,7 @@ class Translate : public GeometricalObject {
 
  public:
   Translate(std::shared_ptr<GeoObject> object, const Vector3& offset) : object(object), offset(offset) {}
-  virtual bool intersect(const Ray& ray, float tMin, float tMax, SInteraction& interaction) const override {
+  virtual bool intersect(const Ray& ray, double tMin, double tMax, SInteraction& interaction) const override {
     const Ray movedRay(ray.origin - offset, ray.direction, ray.getTime());
     if (!object->intersect(movedRay, tMin, tMax, interaction)) return false;
 
@@ -18,7 +18,7 @@ class Translate : public GeometricalObject {
     interaction.setFaceNormal(movedRay, interaction.normal);
     return true;
   }
-  virtual bool computeBoundingBox(float t0, float t1, AABB& outputBox) const override {
+  virtual bool computeBoundingBox(double t0, double t1, AABB& outputBox) const override {
     if (!object->computeBoundingBox(t0, t1, outputBox)) return false;
     outputBox = AABB(outputBox.getMin() + offset, outputBox.getMax() + offset);
     return true;
