+++ src/GeoObjects/MovingSphere.hpp
@@ -8,14 +8,14 @@ class MovingSphere : public GeometricalObject {
  private:
   Point3 center0;
   Point3 center1;
-  float time0;
-  float time1;
-  float radius;
+  double time0;
+  double time1;
+  double radius;
   std::shared_ptr<Material> material;
 
  public:
   MovingSphere() {}
-  MovingSphere(Point3 center0, Point3 center1, float time0, float time1, float radius,
+  MovingSphere(Point3 center0, Point3 center1, double time0, double time1, double radius,
                std::shared_ptr<Material> material) :
       center0(center0),
       center1(center1),
@@ -24,7 +24,7 @@ class MovingSphere : public GeometricalObject {
       radius(radius),
       material(material) {}
 
-  virtual bool intersect(const Ray& ray, float tMin, float tMax, SInteraction& interaction) const override {
+  virtual bool intersect(const Ray& ray, double tMin, double tMax, SInteraction& interaction) const override {
     const Vector3 oc = ray.origin - centerAt(ray.getTime());
     const auto a = ray.direction.magnitudeSquared();
     const auto halfB = dot(oc, ray.direction);
@@ -58,7 +58,7 @@ class MovingSphere : public GeometricalObject {
     return false;
   }
 
-  virtual bool computeBoundingBox(float t0, float t1, AABB& outputBox) const override {
+  virtual bool computeBoundingBox(double t0, double t1, AABB& outputBox) const override {
     const Point3 minPoint = centerAt(t0) - Vector3(radius, radius, radius);
     const Point3 maxPoint = centerAt(t0) + Vector3(radius, radius, radius);
     const auto initBox = AABB(minPoint, maxPoint);
@@ -69,7 +69,7 @@ class MovingSphere : public GeometricalObject {
 	
   std::shared_ptr<Material> getMaterial() const override { return material; }
 
-  Point3 centerAt(float time) const { return center0 + ((time - time0) / (time1 - time0)) * (center1 - center0); }
+  Point3 centerAt(double time) const { return center0 + ((time - time0) / (time1 - time0)) * (center1 - center0); }
 };
 
 #endif
