+++ src/Materials/Lambertian.hpp
@@ -32,8 +32,8 @@ class Lambertian : public Material {
   }
   Color emit(const UV& uv, const Point3& point) const override { return albedo->lookup(uv, point); }
 
-  float brdf(const Vector3& wi, const Vector3& normal, const Vector3& wo) const override { return 1.0 / Math::pi; }
-  float pdf(const Vector3& wi, const Vector3& normal, const Vector3& wo) const override {
+  double brdf(const Vector3& wi, const Vector3& normal, const Vector3& wo) const override { return 1.0 / Math::pi; }
+  double pdf(const Vector3& wi, const Vector3& normal, const Vector3& wo) const override {
     return std::abs(dot(wo, normal)) / Math::pi;
   }
 };
